refactor shoppingcart
