+++ webproject/shoppingcart/context_processor.py
